+++ changelog.md
@@ -38,6 +38,7 @@ This version is not yet released. If you are reading this on the website, then t
   - You can read about the new behavior [here](https://uiua.org/tutorial/codetactility#line-manipulation)
 - Add unicode escape sequences for entering arbitrary characters
   - A double backslash followed by a unicode hex number will format to the corresponding glyph
+- Add the [`timezone`](https://uiua.org/docs/timezone) function, which gets the local timezone offset in hours
 - Add a `Lorem` constant, which contains the Lorem Ipsum text
 - Add several color constants
 - Add pride flag constants